refactor($Presentation) : 계층 구조에 맞는 패키지 명 변경
view -> presentation으로 변경
